requirements.txt
anthropic
openai
python-dotenv
pytesseract
Pillow
ffmpeg-python
requests
